refactor code structure to include shared types into types.rs and avoid creating too many modules
