--- src/chunk/mod.rs
@@ -1,2 +1,0 @@
-mod android;
-mod sample;
