--- src/clientsdk/mod.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ClientSDK {
-    pub name: String,
-    pub version: String,
-}
